% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earnings_hours}
\alias{earnings_hours}
\title{Aggregate Hours and Earnings Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 57 rows and 9 columns.
}
\source{
\href{https://www.ons.gov.uk/employmentandlabourmarket/peopleinwork/earningsandworkinghours/datasets/ashe1997to2015selectedestimates}{Annual Survey of Hours and Earnings time series of selected estimates} 26 October 2022 release.
}
\usage{
earnings_hours
}
\description{
Time series of aggregate data on mean and median hourly earnings (excluding overtime) and weekly paid hours. Also median weekly
and annual earnings. Produced by the Office for National Statistics.
}
\keyword{datasets}
